code cleanup - removed extra custom modules and integrated them to main script
